Defined a form component for new notes. #3
