+++ client/index.js
@@ -1,39 +1,22 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
+import Notes from '../client/notes'
+import Form from '../client/form'
 
 const $app = document.querySelector('#app')
 
-function Note(props) {
-  const { note } = props
-  return (
-      <div className='ui card blue label'>
-        <div className='content'>
-          <div className='header'>{note.id}</div>
-          <div className='meta'>Notes</div>
-          <div className='description'>{note.note_text}</div>
-        </div>
-      </div>
-  )
-}
-
-class Notes extends React.Component {
-  constructor(props) {
-    super(props)
-    this.state = {notes: []}
-  }
-
-  async componentDidMount() {
-      const response = await fetch('/notes')
-      const notes = await response.json()
-      this.setState({ notes })
-    }
+class Main extends React.Component {
   render() {
     return (
       <div>
-        {this.state.notes.map((note, i) => <Note key={i} note={note} />)}
+        <div>
+          <Notes />
+          <br></br>
+          <Form />
+        </div>
       </div>
     )
   }
 }
 
-ReactDOM.render(<Notes />, $app)
+ReactDOM.render(<Main />, $app)
